requirements.txt
playwright
python-telegram-bot
requests
python-dotenv
requests-oauthlib
openai
